+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/ImpTest.hs
@@ -1662,6 +1662,7 @@ tryTxsInBlock' txs finalState blockIssuer = do
         , tbhBHash = hashBlockBody blockBody
         , tbhSlot = slotNo
         , tbhVersionInfo = BlockHeaderVersionInfo (getVersion32 curMajor) curMinor
+        , tbhEbRefsAnn = SNothing
         }
     block = Block {blockHeader, blockBody}